更新代码

+++ uninstall.php
@@ -13,4 +13,6 @@ if( !defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	exit;
 }
 
-delete_option("editormd");
+if ( get_option( 'editormd' ) ) {
+	delete_option( 'editormd' );
+}
